CustomButton: remove unnecessary code; style testing is handled by cypress (#13)
Testing the Custom Button style is better handled by cypress: checking how a component looks in the browser is cypress' motive and thus we test this over there.

The way the testing-library was executing this test was by literally checking if there's a style applied to the button; remember that testing-library relies on the creation of a virtual DOM tree and thus, no actual styling is applied. Cypress can actually check how the button looks to the user since it is directly accessing the DOM generated in the browser.